[refactor] 리뷰 필터링 기능 리팩토링
- CategoryBar 컴포넌트를 shared 레이어에서 features 레이어로 이동. (카테고리를 선택해 리뷰를 필터링하는 '기능'이기 때문)
- 카테고리 선택 로직을 useSelectCategory 커스텀 훅으로 분리. (이후 게시글 모음 페이지에서 재사용)
- 카테고리 리스트를 별도 상수 파일로 분리. (categoryList.ts)
