+++ src/features/review-filtering/index.ts
@@ -1,0 +1,2 @@
+export {default as CategoryBar} from './ui/CategoryBar';
+export {useSelectCategory} from './lib/useSelectCategory';
