+++ src/features/review-filtering/ui/CategoryBar.tsx
@@ -1,28 +1,13 @@
 import {BestReviewCategory} from '@/entities/reviews/model/types';
 import {Button} from '@/shared/shadcnComponent/ui/button';
+import CATEGORY_LIST from '../consts/categoryList';
 
 type Props = {
   selectedCategory: BestReviewCategory;
   onSelectCategory: (category: BestReviewCategory) => void;
 };
 
-type Category = {
-  id: BestReviewCategory;
-  category: string;
-};
-
 export default function CategoryBar({selectedCategory, onSelectCategory}: Props) {
-  const CATEGORY_LIST: Category[] = [
-    {id: 'all', category: '전체'},
-    {id: 'car', category: '자동차'},
-    {id: 'food', category: '음식'},
-    {id: 'company', category: '회사'},
-    {id: 'cosmetic', category: '화장품'},
-    {id: 'cafe', category: '카페'},
-    {id: 'devices1', category: '전자제품'},
-    {id: 'clothes', category: '의류'},
-  ];
-
   return (
     <nav className="w-full px-2 mb-14">
       <ul className="grid grid-cols-[repeat(auto-fit,minmax(100px,1fr))] text-center gap-2 bg-white rounded-3xl p-3">
