+++ src/shared/ui/components/index.ts
@@ -1,2 +1,1 @@
-export {default as CategoryBar} from './CategoryBar';
 export {default as Toast, type ToastProps} from './Toast';
